@rparrett 2025-03-10
I think the minor differences can be attributed to swash `0.2.0`, but it's difficult to follow the changes to that library and its dependencies.