@rparrett 2025-03-10
Not seeing anything concerning in there, just the same minor differences noticed by the testbed.

However, it seems that `cosmic-text` is failing to build for android.