@mnmaita 2025-03-10
https://github.com/pop-os/cosmic-text/issues/349 looks like someone had the same issue with the previous version too?