@rparrett 2025-03-11
Reverting https://github.com/pop-os/cosmic-text/commit/829a59bd9567b1ce1dd38c2fd527cf9ff838625b seems to fix it. I guess android targets used to take the "no-op" path in `other.rs` and that's not happening anymore.

I have a branch that works but I'm not 100% confident in. Will open a PR for discussion.