@rparrett 2025-03-11
https://github.com/pop-os/cosmic-text/pull/365