@alice-i-cecile 2025-03-11
Bumping from 0.17; we shouldn't wait on a fix IMO. This is not essential to ship immediately.