@rparrett 2025-03-11
`cosmic-text` `0.13.2` was released which should fix the Android issue.